requirements.txt
opencv_python==4.4.0.46
numpy==1.19.2
pytube==9.7.2
streamlit==0.69.2
Pillow==8.0.1
